MarketVisionAI/backend/portfolio/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
google-generativeai==0.8.5